Change name of the package to conform to the MavenCentral
