+++ hpotextmining-core/pom.xml
@@ -5,16 +5,15 @@
     <modelVersion>4.0.0</modelVersion>
 
     <parent>
-        <artifactId>HpoTextMining</artifactId>
-        <groupId>org.monarchinitiative</groupId>
-        <version>0.1.1-SNAPSHOT</version>
+        <artifactId>hpotextmining</artifactId>
+        <groupId>com.github.monarch-initiative</groupId>
+        <version>0.1.0</version>
     </parent>
 
     <artifactId>hpotextmining-core</artifactId>
     <packaging>jar</packaging>
-    <version>${parent.version}</version>
 
-    <description>Use this module to perform text mining for HPO terms from publication.</description>
+    <description>Use this module to perform text mining for HPO terms</description>
 
     <properties>
         <log4j.version>2.7</log4j.version>
