+++ hpotextmining-core/src/test/java/com/github/monarchinitiative/hpotextmining/model/PhenotypeTermTest.java
@@ -1,0 +1,45 @@
+package com.github.monarchinitiative.hpotextmining.model;
+
+import org.junit.Before;
+import org.junit.Test;
+
+import static org.junit.Assert.assertEquals;
+import static org.junit.Assert.assertNotEquals;
+
+/**
+ * Tests of the {@link PhenotypeTerm} class.
+ *
+ * @author <a href="mailto:daniel.danis@jax.org">Daniel Danis</a>
+ * @version 0.1.0
+ * @since 0.1
+ */
+public class PhenotypeTermTest {
+
+    private PhenotypeTerm hypertension1, hypertension2, hypertension3, hepatosplenomegaly;
+
+
+    @Before
+    public void setUp() throws Exception {
+        hypertension1 = new PhenotypeTerm("HP:0000822", "Hypertension", "The presence of chronic increased pressure in the " +
+                "systemic arterial system.", true);
+        hypertension2 = new PhenotypeTerm("HP:0000822", "Hypertension", "The presence of chronic increased " +
+                "pressure in the systemic arterial system.", true);
+        hypertension3 = new PhenotypeTerm("HP:0000822", "Hypertension", "The presence of chronic increased " +
+                "pressure in the systemic arterial system.", false);
+        hepatosplenomegaly = new PhenotypeTerm("HP:0001433", "Hepatosplenomegaly", "Simultaneous enlargement of the liver and spleen.", true);
+    }
+
+
+    /**
+     * Test correct behaviour of the {@link PhenotypeTerm#equals(Object)} method.
+     *
+     * @throws Exception bla
+     */
+    @Test
+    public void testEqualsTo() throws Exception {
+        assertEquals(hypertension1, hypertension2);
+        assertNotEquals(hypertension1, hypertension3);
+        assertNotEquals(hypertension1, hepatosplenomegaly);
+
+    }
+}
