+++ hpotextmining-demo/pom.xml
@@ -2,15 +2,16 @@
 <project xmlns="http://maven.apache.org/POM/4.0.0"
          xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
          xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
-    <parent>
-        <artifactId>HpoTextMining</artifactId>
-        <groupId>org.monarchinitiative</groupId>
-        <version>0.1.1-SNAPSHOT</version>
-    </parent>
     <modelVersion>4.0.0</modelVersion>
 
+    <parent>
+        <artifactId>hpotextmining</artifactId>
+        <groupId>com.github.monarch-initiative</groupId>
+        <version>0.1.0</version>
+    </parent>
+
     <artifactId>hpotextmining-demo</artifactId>
-    <description>Example usage of `hpotextmining-core` module in a JavaFX App.</description>
+    <description>Example usage of `hpotextmining-core` module in JavaFX App</description>
 
     <properties>
         <spring.version>4.3.11.RELEASE</spring.version>
@@ -20,6 +21,16 @@
         <finalName>${project.artifactId}-${project.version}</finalName>
 
         <plugins>
+            <plugin>
+                <groupId>org.apache.maven.plugins</groupId>
+                <artifactId>maven-compiler-plugin</artifactId>
+                <version>2.3.2</version>
+                <configuration>
+                    <source>${jdk.version}</source>
+                    <target>${jdk.version}</target>
+                </configuration>
+            </plugin>
+
             <plugin>
                 <groupId>org.apache.maven.plugins</groupId>
                 <artifactId>maven-dependency-plugin</artifactId>
@@ -84,7 +95,7 @@
     </build>
     <dependencies>
         <dependency>
-            <groupId>org.monarchinitiative</groupId>
+            <groupId>com.github.monarch-initiative</groupId>
             <artifactId>hpotextmining-core</artifactId>
             <version>0.1.0</version>
         </dependency>
